+++ madprops/serializers.py
@@ -4,8 +4,11 @@ from django.db.models import ForeignKey
 from django.utils.functional import cached_property
 from django.utils.translation import ugettext as _
 from rest_framework.exceptions import ValidationError
-from rest_framework.serializers import (BaseSerializer, ListSerializer,
-                                        ModelSerializer)
+from rest_framework.serializers import (
+    BaseSerializer,
+    ListSerializer,
+    ModelSerializer,
+)
 
 
 class PropertySerializerOptions(object):
@@ -133,8 +136,7 @@ class PropertySerializer(ModelSerializer):
 
         # If it already exists - update it's value. Otherwise - create a new
         # property.
-        existing_props = self.Meta.model.objects.filter(**filters)
-        prop = existing_props[0] if existing_props else None
+        prop = self.Meta.model.objects.filter(**filters).first()
 
         if prop_name in self.opts.read_only_props:
             return prop
@@ -162,7 +164,7 @@ class PropertySerializer(ModelSerializer):
         {<prop_name>: <prop_value>} ->
             {'name': <prop_name>, 'value': <prop_value>}
         """
-        prop_name, prop_value = data.items()[0]
+        prop_name, prop_value = next(iter(data.items()))
         return {'name': prop_name, 'value': prop_value}
 
     def _add_parent_obj_field(self, data):